The following is tooken from http://projects.zoho.com/portal/nopcommerce#forums/307368000000011021/307368000000045023
I don't think people who implement configurable IPlugins should have to create an entirely different class that implements ISettings to get and save settings.

I think we should support services (and inherently plugins) that can implement ISettings on the same class.

pubic class PaypalPayementProvider : IPaymentPlugin, ISettings

My goal, however, is too create an IConfigurablePlugin that implements IPlugin and ISettings. Then, in the IPluginFinder class, we can inject the setting values into a plugin instance, making getting and saving values very transparent.

I have refactored DependencyRegister and SettingsSource and also created a new SettingsModule (autofac) module. This is how I set things up. I made it so that the container will return all types implement ISettings, even ones that are configured to run as services.

public class ClassImplementingISettingsButNotRegisteredWithContainer

public class ClassImplementingISettingsButIsRegisteredToBeResolvedByTheContainer

both of the classes will be resolved by either using the container, or using ContainerManager.ResolveUnregistered.

But how are the settings value's getting loaded?

The SettingsModule basically attaches events to all registrations that are services for an implementation of ISettings. If it is, it attaches an OnActivated event that is used to touch the instance before it is used to satisfy a request. This is where we can load our settings.

All this required me to make one more method on IConfigurationProvider called LoadInto(TSettings) that injects settings into an already instantiated (via Resolve or ResolveUnknown) services.